Add stub for http upgrade handling. Rename nodej to nodyn and fix.
HTTP upgrade handling is tricky. In node.js, you are required to handle
the HTTP Upgrade header yourself and send an appropriate response. If
you don't, the connection is closed. Vert.x, on the other hand hides all
of this from you to the extent that I don't see how it's possible for me
to implement this part of the node.js http API cleanly without a lot of
wrapper code for the socket.  We'll see...

In Node, you listen for the 'upgrade' event on both the client and
server side. The client explicitly sends an upgrade header, while the
server explicitly looks for one. When an upgrade is requested, the
server is provided a handle on the actual request, the underlying
socket, and possibly the first packet of the upgraded stream. Ugh.

http://nodejs.org/api/http.html#http_event_upgrade_1