Rework project into two modules: nodyn-verticle and nodyn-standalone.
This allows for a nice command line REPL for using nodyn informally,
testing out things, etc. The nodyn REPL is essentially the DynJS REPL
with Vert.x embedded, and mod-lang-js and nodyn-verticle pre-loaded.
Using the nodyn repl does not allow for usage of vertx-platform,
however, so the vertx/container module will not function.

Example:

    ~/s/v/n/nodyn-standalone git:multi-module ❯❯❯ ./bin/nodyn --console
    nodyn> vertxHttp = require('vertx/http')
    org.dynjs.runtime.DynObject@40ae49d5
    nodyn> nodeHttp = require('http')
    org.dynjs.runtime.DynObject@31dbfb7a
    nodyn>