Add specRunner.js as top level spec responsible for executing others.
This is due to a change in the latest vert.x and its upgrade to netty
version 4.0.19. It is now required to call vertx.stop() if you want the
test process to exit cleanly - specRunner.js handles this. If you want
to run a single spec, you can still use `-Dtest.pattern=...` but you'll
need to kill the process with `CTL-C` when the tests complete.