Remove Vert.x JS language bits from dgram.js.
This completes the removal of the Vert.x Javascript API from the
project. There is still a dependency on lang-dynjs specifically because
NodeJSVerticleFactory still extends DynJSVerticleFactory. But this begs
the question, wtf do we even still have a NodeJSVerticleFactory, as I
believe with our new embedded approach it is no longer needed. Next!