Implement datagram send and recieve bindings.
There are still some questions about this. First of all, the bind call
to netty does not block, so the implementation is racy. In fact, if you
run the tests multiple times you'll certainly see it hang. I've also had
to do some ByteBuf copying which I think may be non-performant and I'd
like to figure out a way to avoid it if possible.