Greatly simplify the jasmine test architecture.
1. Removed dependency on jasmine-jvm
2. Added a more complete test reporter with nice color output and which
can recognize when a test times and, and report that fact on the
terminal. The reporter comes from
https://github.com/larrymyers/jasmine-reporters
3. Made TestRunner.java startup script much simpler. It now only passes
the test pattern as a string to specRunner.js' run() function.
4. Scanning the file system for which specs to run is done in a simple
javascript file scanner.js

This all started because I was testing fs.watch() which seemed to be
timing out. Un-`xit` that spec and you can easily see the timeout
message dislayed by the terminal reporter.