Provide a super special magical .jar require() extension.
If the require('mything') points to a .jar, it will
add it to the classpath.

If it additionally contains a root-level entry named
'node.extension', and that file contains a single non-blank
and non-comment line specifying a classname, and that
class contains a method matching this name and signature:

  public static Object require()

It will invoke that static method, and return the return
value as the equivalent of module.exports for that require()
invocation.