Add native node.js test for zlib objects.
This test has the capability to test all compression/decompression
combinations with a text file, an empty file, and a binary file.
Currently we are only testing Inflate and Deflate. Changes, still
undisovered, are required for Gzip/Gunzip and the generic Unzip.

This test can also be altered to run with all possible combinations of
compression levels, windowBits, memLevel and strategy. To do this,
process.env.PUMMEL should be true.